+++ src/main/java/swimming/pool/infra/mybatis/CustomerMapper.java
@@ -1,11 +1,11 @@
 package swimming.pool.infra.mybatis;
 
 import org.apache.ibatis.annotations.Mapper;
-import swimming.pool.domain.customer.Customer;
+import swimming.pool.infra.mybatis.mappermodel.CustomerInfoDao;
 
 @Mapper
 public interface CustomerMapper {
 
-  void insert(Customer customer);
+  void insert(CustomerInfoDao customer);
 
 }
