+++ src/main/java/swimming/pool/infra/repository/CustomerRepository.java
@@ -3,6 +3,7 @@ package swimming.pool.infra.repository;
 import org.springframework.stereotype.Repository;
 import swimming.pool.infra.mybatis.CustomerMapper;
 import swimming.pool.domain.customer.Customer;
+import swimming.pool.infra.mybatis.mappermodel.CustomerInfoDao;
 
 @Repository
 public class CustomerRepository {
@@ -14,6 +15,7 @@ public class CustomerRepository {
   }
 
   public void save(Customer lee) {
-    customerMapper.insert(lee);
+      CustomerInfoDao dao = new CustomerInfoDao(lee.getName(), lee.getAge());
+    customerMapper.insert(dao);
   }
 }
