+++ src/main/resources/mybatis/Customer.xml
@@ -4,8 +4,8 @@
   "http://mybatis.org/dtd/mybatis-3-mapper.dtd">
 
 <mapper namespace="swimming.pool.infra.mybatis.CustomerMapper">
-  <insert id="insert" parameterType="swimming.pool.domain.customer.Customer"
-    useGeneratedKeys="true" keyProperty="id">
+  <insert id="insert" parameterType="swimming.pool.infra.mybatis.mappermodel.CustomerInfoDao"
+    useGeneratedKeys="true" keyProperty="customerId">
     INSERT
     INTO customer(name, age)
     VALUES (#{name}, #{age})
